Correctly determine if we are creating a session or not when calling rpc.php.
Need to determine type of server before calling appInit() and also need to let
Horde_Controller_Request_Http know we don't want a session.  This will probably
be taken care of when we use a full-fledged controller here, but for now we need to
pass the session_control value to the Request const'r.