Add Request Body Stream to interface
I noticed Horde_Controller currently lacks any access to the body of a http request like post or put, unless it's form-encoded. In this case, getPostVars works well
For the REST usecase, I would want to post/put a json object string containing details.

Another use case would be handling some bulk upload like a file (owncloudish) or a media stream.

For the REST use case, just slurping stdin and returning it as a string would be sufficient.
For the latter, it would be more convenient to expose the stream resource - both memory and performance wise.
Horde_Stream has a __toString() to implicitly return a string where one is expected.